+++ opencc_purepy/__init__.py
@@ -4,3 +4,4 @@
 # January, 2025
 ##########################################################
 from .core import OpenCC, OpenccConfig
+from .dict_slot import DictSlot, DictSlotLike